+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 - TT-1993 - Use DIT sectors
 - TT-2024 - Expose expires timestamp to save for later email
 - TT-2019 - Use DIT countries and territories
+- TT-2016 - Add autocomplete multiselect
 
 ### Bug fixes